backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-migrate==4.0.7
SQLAlchemy==2.0.35
psycopg[binary]==3.2.9
redis==5.0.7
pydantic==2.8.2
pydantic-settings==2.4.0
python-slugify==8.0.4
marshmallow==3.22.0
python-dotenv==1.0.1
